Refactored menu package. Now Menu is a PersistentObject

+++ community/logicaldoc/trunk/logicaldoc-core/src/main/java/com/logicaldoc/core/searchengine/Result.java
@@ -48,7 +48,7 @@ public interface Result {
 
 	public void setSourceDate(Date date);
 	
-	public void setSize(int sze);
+	public void setSize(long sze);
 
 	public void setType(String typ);
